Create Feature Extraction Dataset
Why the feature extraction model? In the final object tracking algorithm, the model should be able to identify the entry and exit (basically presence) of the vehicle in the drone's video stream. This can be achieved using the saved training weights of this feature extraction model. 
Dataset Preparation:
Snapshots of target objects from various video sources across the internet from their entry to exit 
21 different cars were selected for dataset preparation